+++ hs_core/management/commands/solr_facets.py
@@ -4,7 +4,6 @@ It is used for debugging the faceting system.
 """
 
 from django.core.management.base import BaseCommand
-from hs_core.discovery_form import FIELDS_TO_SHOW
 from haystack.query import SearchQuerySet
 from pprint import pprint
