--- hs_core/templates/search/indexes/hs_core/baseresource_json.txt
@@ -1,27 +1,0 @@
-{% load hydroshare_tags %}{ 
- "short_id":"{% if object.short_id %}{{ object.short_id }}{% endif %}", 
- "doi":"{% if object.doi %}{{ object.doi }}{% endif %}", 
- "title":"{% if object.metadata.title.value %}{{ object.metadata.title.value }}{% endif %}", 
- "publisher":"{% if object.metadata.publisher.name %}{{ object.metadata.publisher.name }}{% endif %}",
- "resource_type":"{% if object.resource_type %}{{ object.resource_type }}{% endif %}", 
- "creators":[
-{% for creator in object.metadata.creators.all %}  { 
-   "name":"{% if creator.name %}{{ creator.normalize_human_name }}{% endif %}",
-   "organization":"{% if creator.organization %}{{ creator.organization }}{% endif %}"
-  }{% if not forloop.last %},{% endif %}{% endfor %}
- ],
- "contributors":[{% for contributor in object.metadata.contributors.all %}  { 
-  "name":"{% if contributor.name %}{{ contributor.name|normalize_human_name }}{% endif %}",
-  "organization":"{% if contributor.organization %}{{ contributor.organization }}{% endif %}"
-  }{% if not forloop.last %},{% endif %}{% endfor %}
- ], 
- "subjects":["{{ object.metadata.subjects.all|join:'","' }}"] , 
- "owners":[{% for owner in object.raccess.owners.all %} 
-  {
-   "username":"{{ owner.username }}", 
-   "first": "{{ owner.first_name }}", 
-   "last":"{{owner.last_name}}"
-  }{% if not forloop.last %},{% endif %}{% endfor %}
- ], 
- "first_author":"{{ object.first_creator }}"
-}
