--- hs_core/templates/search/indexes/hs_core/baseresource_text.txt
@@ -1,22 +1,0 @@
-{% load hydroshare_tags %} 
-{% if object.short_id %} {{ object.short_id }} {% endif %} 
-{% if object.doi %} {{ object.doi }} {% endif %} 
-{% if object.metadata.title.value %} {{ object.metadata.title.value }} {% endif %} 
-{% if object.metadata.description %} {{ object.metadata.description }} {% endif %} 
-{% if object.metadata.publisher.name %} {{ object.metadata.publisher.name }} {% endif %} 
-{% if object.resource_type %} {{ object.resource_type }} {% endif %} 
-{% for creator in object.metadata.creators.all %}
-    {% if creator.name %} {{ creator.name }} {{ creator.normalize_human_name }} {% endif %} 
-    {% if creator.organization %} {{ creator.organization }} {% endif %} 
-{% endfor %}
-{% for contributor in object.metadata.contributors.all %}
-    {% if contributor.name %} {{ contributor.name }} {{ contributor.name|normalize_human_name
-    {% if contributor.organization %} {{ contributor.organization }} {% endif %} 
-{% endif %} 
-{% endfor %}
-{% for subject in object.metadata.subjects.all %}
-    {% if subject %} {{ subject }} {% endif %} 
-{% endfor %}
-{% for owner in object.raccess.owners.all %}
-    {{ owner.username }} {{ owner.first_name }} {{owner.last_name}}, {{owner.first_name}}
-{% endfor %}
