+++ hydroshare/urls.py
@@ -10,7 +10,6 @@ from mezzanine.pages.views import page
 
 from autocomplete_light import shortcuts as autocomplete_light
 
-from hs_core.views.discovery_view import DiscoveryView
 from hs_core.views.discovery_json_view import DiscoveryJsonView
 from hs_core.views.oauth2_view import GroupAuthorizationView
 from hs_sitemap.views import sitemap
@@ -21,7 +20,6 @@ from hs_app_timeseries import views as hs_ts_views
 import hs_communities.views.communities
 from theme.views import delete_resource_comment
 from hs_discover.views import SearchView, SearchAPI
-from django.views.decorators.cache import cache_page
 
 autocomplete_light.autodiscover()
 admin.autodiscover()
@@ -72,7 +70,6 @@ urlpatterns = i18n_patterns(
     url(r'^verify/(?P<token>[0-9a-zA-Z:_\-]*)/', hs_core_views.verify),
     url(r'^django_irods/', include('django_irods.urls')),
     url(r'^autocomplete/', include('autocomplete_light.urls')),
-    url(r'^discover/$', DiscoveryView.as_view(), name='haystack_search'),
     url(r'^discoverapi/$', SearchAPI.as_view(), name='DiscoverAPI'),
     url(r'^search/$', SearchView.as_view(), name='Discover'),
     url(r'^topics/$', hs_communities.views.communities.TopicsView.as_view(), name='topics'),
