+++ theme/templates/newbase.html
@@ -90,18 +90,6 @@
 
     {% load inplace_edit %}
     {% inplace_static %}
-    <!-- HTML5 shim and Respond.js IE8 support of HTML5 elements and media queries -->
-    <!--[if lt IE 9]>
-{#    <script src="https://oss.maxcdn.com/libs/html5shiv/3.7.0/html5shiv.js"></script>#}
-{#    <script src="https://oss.maxcdn.com/libs/respond.js/1.3.0/respond.min.js"></script>#}
-    <![endif]-->
-    <!--[if IE]>
-{#    <style>#}
-{#        .showcase {#}
-{#            zoom: 0.6;#}
-{#        }#}
-{#    </style>#}
-    <![endif]-->
 
     {% block extra_head %}{% endblock %}
     {% if not page|is_debug %}
